feat: rename theme and update README
Rename the theme and change some references to the upstream author. Keep some, for example in the copyright declaration.

Kept example site config and content as-is, since it's just an example.

Declare in the README that this theme exists solely for my own blog, and any other use is unsupported.